requirements.txt
streamlit~=1.32.1
openai~=1.14.1
requests~=2.31.0
openai-whisper